refactor(data-provider): separate concerns from `schemas` into `parsers`, `config`, and fix imports elsewhere
