+++ packages/data-provider/src/config.ts
@@ -1,0 +1,180 @@
+import { z } from 'zod';
+import { EModelEndpoint } from './schemas';
+
+export const endpointSchema = z.object({
+  name: z.string(),
+  apiKey: z.string(),
+  baseURL: z.string(),
+  models: z.object({
+    default: z.array(z.string()).min(1),
+    fetch: z.boolean().optional(),
+  }),
+  titleConvo: z.boolean().optional(),
+  titleMethod: z.union([z.literal('completion'), z.literal('functions')]).optional(),
+  titleModel: z.string().optional(),
+  summarize: z.boolean().optional(),
+  summaryModel: z.string().optional(),
+  forcePrompt: z.boolean().optional(),
+  modelDisplayLabel: z.string().optional(),
+});
+
+export const configSchema = z.object({
+  version: z.string(),
+  cache: z.boolean(),
+  endpoints: z.object({
+    custom: z.array(endpointSchema.partial()),
+  }),
+});
+
+export enum KnownEndpoints {
+  mistral = 'mistral',
+  openrouter = 'openrouter',
+}
+
+export const defaultEndpoints: EModelEndpoint[] = [
+  EModelEndpoint.openAI,
+  EModelEndpoint.assistant,
+  EModelEndpoint.azureOpenAI,
+  EModelEndpoint.bingAI,
+  EModelEndpoint.chatGPTBrowser,
+  EModelEndpoint.gptPlugins,
+  EModelEndpoint.google,
+  EModelEndpoint.anthropic,
+  EModelEndpoint.custom,
+];
+
+export const alternateName = {
+  [EModelEndpoint.openAI]: 'OpenAI',
+  [EModelEndpoint.assistant]: 'Assistants',
+  [EModelEndpoint.azureOpenAI]: 'Azure OpenAI',
+  [EModelEndpoint.bingAI]: 'Bing',
+  [EModelEndpoint.chatGPTBrowser]: 'ChatGPT',
+  [EModelEndpoint.gptPlugins]: 'Plugins',
+  [EModelEndpoint.google]: 'Google',
+  [EModelEndpoint.anthropic]: 'Anthropic',
+  [EModelEndpoint.custom]: 'Custom',
+};
+
+export const defaultModels = {
+  [EModelEndpoint.google]: [
+    'gemini-pro',
+    'gemini-pro-vision',
+    'chat-bison',
+    'chat-bison-32k',
+    'codechat-bison',
+    'codechat-bison-32k',
+    'text-bison',
+    'text-bison-32k',
+    'text-unicorn',
+    'code-gecko',
+    'code-bison',
+    'code-bison-32k',
+  ],
+  [EModelEndpoint.anthropic]: [
+    'claude-2.1',
+    'claude-2',
+    'claude-1.2',
+    'claude-1',
+    'claude-1-100k',
+    'claude-instant-1',
+    'claude-instant-1-100k',
+  ],
+  [EModelEndpoint.openAI]: [
+    'gpt-3.5-turbo-16k-0613',
+    'gpt-3.5-turbo-16k',
+    'gpt-4-1106-preview',
+    'gpt-3.5-turbo',
+    'gpt-3.5-turbo-1106',
+    'gpt-4-vision-preview',
+    'gpt-4',
+    'gpt-3.5-turbo-instruct-0914',
+    'gpt-3.5-turbo-0613',
+    'gpt-3.5-turbo-0301',
+    'gpt-3.5-turbo-instruct',
+    'gpt-4-0613',
+    'text-davinci-003',
+    'gpt-4-0314',
+  ],
+};
+
+export const EndpointURLs: { [key in EModelEndpoint]: string } = {
+  [EModelEndpoint.openAI]: `/api/ask/${EModelEndpoint.openAI}`,
+  [EModelEndpoint.bingAI]: `/api/ask/${EModelEndpoint.bingAI}`,
+  [EModelEndpoint.google]: `/api/ask/${EModelEndpoint.google}`,
+  [EModelEndpoint.custom]: `/api/ask/${EModelEndpoint.custom}`,
+  [EModelEndpoint.anthropic]: `/api/ask/${EModelEndpoint.anthropic}`,
+  [EModelEndpoint.gptPlugins]: `/api/ask/${EModelEndpoint.gptPlugins}`,
+  [EModelEndpoint.azureOpenAI]: `/api/ask/${EModelEndpoint.azureOpenAI}`,
+  [EModelEndpoint.chatGPTBrowser]: `/api/ask/${EModelEndpoint.chatGPTBrowser}`,
+  [EModelEndpoint.assistant]: '/api/assistants/chat',
+};
+
+export const modularEndpoints = new Set<EModelEndpoint | string>([
+  EModelEndpoint.gptPlugins,
+  EModelEndpoint.anthropic,
+  EModelEndpoint.google,
+  EModelEndpoint.openAI,
+  EModelEndpoint.azureOpenAI,
+  EModelEndpoint.custom,
+]);
+
+export const supportsFiles = {
+  [EModelEndpoint.openAI]: true,
+  [EModelEndpoint.google]: true,
+  [EModelEndpoint.assistant]: true,
+  [EModelEndpoint.azureOpenAI]: true,
+  [EModelEndpoint.custom]: true,
+};
+
+export const supportsBalanceCheck = {
+  [EModelEndpoint.openAI]: true,
+  [EModelEndpoint.azureOpenAI]: true,
+  [EModelEndpoint.gptPlugins]: true,
+  [EModelEndpoint.custom]: true,
+};
+
+export const visionModels = ['gpt-4-vision', 'llava-13b', 'gemini-pro-vision'];
+
+/**
+ * Enum for cache keys.
+ */
+export enum CacheKeys {
+  /**
+   * Key for the config store namespace.
+   */
+  CONFIG_STORE = 'configStore',
+  /**
+   * Key for the plugins cache.
+   */
+  PLUGINS = 'plugins',
+  /**
+   * Key for the model config cache.
+   */
+  MODELS_CONFIG = 'modelsConfig',
+  /**
+   * Key for the default endpoint config cache.
+   */
+  ENDPOINT_CONFIG = 'endpointsConfig',
+  /**
+   * Key for the custom config cache.
+   */
+  CUSTOM_CONFIG = 'customConfig',
+  /**
+   * Key for the override config cache.
+   */
+  OVERRIDE_CONFIG = 'overrideConfig',
+}
+
+/**
+ * Enum for authentication keys.
+ */
+export enum AuthKeys {
+  /**
+   * Key for the Service Account to use Vertex AI.
+   */
+  GOOGLE_SERVICE_KEY = 'GOOGLE_SERVICE_KEY',
+  /**
+   * API key to use Google Generative AI.
+   */
+  GOOGLE_API_KEY = 'GOOGLE_API_KEY',
+}
